+++ progress.c
@@ -27,7 +27,7 @@ extern int am_server;
 extern int flist_eof;
 extern int quiet;
 extern int need_unsorted_flist;
-extern int output_needs_newline;
+extern ROLE_TLS int output_needs_newline;
 extern int stdout_format_has_i;
 extern struct stats stats;
 extern struct file_list *cur_flist;